Implement ldv_krealloc_* using realloc
Removes one of the uses of ldv_malloc_unknown_size, which is implemented
using __VERIFIER_nondet_pointer. The size is known (passed in as
parameter), and `realloc` much more closely resembles the semantics.
This is in preparation of removing uses of __VERIFIER_nondet_pointer
(see #767).

The changes were done using the following command:
```
for f in c/ldv-*/*.[ci]
do
perl -i -e \
  '$p = 1; $d = 0;
   while(<>) {
     if(/^static void \*ldv_krealloc_\d+\(void const  *\*ldv_func_arg1 , size_t ldv_func_arg2 , ?$/) {
       $p = 2;
       if($d == 0) {
         print "extern void *realloc(void * , size_t  ) ;\n";
         $d = 1;
       }
     }
     if($p == 2) {
       if(/^  tmp = ldv_malloc_unknown_size\(\);/) {
         print "  tmp = realloc(ldv_func_arg1, ldv_func_arg2);\n";
         $p = 1;
         next;
       }
       elsif(/^\}$/) { $p = 1; }
     }
     print;
   }' \
  $f
done
```