Implement ldv___get_free_pages_* using ldv_malloc
Removes one of the uses of ldv_malloc_unknown_size, which is implemented
using __VERIFIER_nondet_pointer. The size is known (the second argument
is the order, which is the 2-logarithm of the number of pages, and a
page usually is 4K), thus using ldv_malloc is perfectly safe. This is in
preparation of removing uses of __VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
for f in c/ldv-*/*.[ci]
do
perl -i -e \
  '$p = 1; $d = 0;
   while(<>) {
     if(/^void \*ldv_malloc\(size_t size \) ;$/) {
       next if($d == 2);
       $d = 1;
     }
     if(/^static unsigned long ldv___get_free_pages_\d+\(gfp_t flags , unsigned int ldv_func_arg2 \) ?$/) {
       $p = 2;
       if($d == 0) {
         print "void *ldv_malloc(size_t size ) ;\n";
         $d = 2;
       }
     }
     if($p == 2) {
       if(/^  tmp = ldv_malloc_unknown_size\(\);/) {
         print "  tmp = ldv_malloc(4096UL * (1 << ldv_func_arg2));\n";
         next;
       }
       elsif(/^  return \(\(unsigned long \)\(\(unsigned int \)\(\(long \)tmp\)\)\);/) {
         print "  return tmp;\n";
         $p = 1;
         next;
       }
       elsif(/^\}$/) { $p = 1; }
     }
     print;
   }' \
  $f
done
```