Specialise kmem_cache_alloc to the desired allocation size
kmem_cache_alloc should allocate (from a cache) memory of the size
specified when the cache was created (via kmem_cache_create). Since none
of the benchmarks provide a definition of kmem_cache_create (or struct
kmem_cache), this information cannot be communicated. Instead, hard-code
the allocation size as used in the specific device driver (and the
particular cache, when multiple caches are used) in kmem_cache_alloc and
use ldv_malloc instead of ldv_malloc_unknown_size.  This is in
preparation of removing uses of __VERIFIER_nondet_pointer (see #767).