Specialise mempool_alloc to the desired allocation size
mempool_alloc should allocate (from a memory pool) memory of the size specified
when the pool was created (via mempool_create). Since none of the benchmarks
provide a definition of mempool_create, this information cannot be
communicated. Instead, hard-code the allocation size as used in the specific
device driver in mempool_alloc and use ldv_malloc instead of
ldv_malloc_unknown_size.  This is in preparation of removing uses of
__VERIFIER_nondet_pointer (see #767).