Implement pci_iomap using ldv_malloc
Removes one of the uses of external_alloc, which is implemented using
__VERIFIER_nondet_pointer. We allocate the amount of memory specified in
the maxlen member, and don't currently handle the "0" case (where the
full BAR should be mapped) at the moment.  This is in preparation of
removing uses of __VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
for f in c/ldv-*/model/*.c
do
perl -i -e \
  '$p = 1; $d = 0;
   while(<>) {
     if(/^\/\/ Function: pci_iomap\s*$/) {
       $p = 2;
     }
     if($p == 2) {
       if(/^void \*external_alloc\(void\);\s*$/) { next; }
       elsif(/^  return \(void \*\)external_alloc\(\);$/) {
         print "  return ldv_malloc(arg2);\n";
         $p = 1;
         next;
       }
       elsif(/^\}$/) { $p = 1; }
     }
     print;
   }' \
  $f
done
```