fix undefined behaviour from pointer comparison after a pointer was freed.
According to ISO-C99 6.2.4(2),
the value of a pointer becomes indeterminate when
the object it points to reaches the end of its lifetime.

The undefined behavior comes from computing "x==y" in this task,
because "y" was freed before, so its value is indeterminate.
The fix uses a temporary variable for storing the corresponding value.