change verdict of memory safety (mem-track).
There is at least one struct 'devobj' in the main function,
whose members are allocated on the heap, but never freed.

This violates the memory property for mem-track,
because when returning from the main function,
the struct itself is cleaned up with the stack frame,
but the heap object is no longer reachable.