VerifyThis benchmarks: avoid overflow in calloc
The man page for calloc says: "The  calloc()  function  allocates memory
for  an array of nmemb elements of size bytes each and returns a pointer
to the allocated memory." It is not further specified what the behaviour
is when nmemb * size exceeds the maximum addressable storage. Thus guard
the call to calloc by an assumption that ensures that nmemb is strictly
less than 2^30 (this is a 32-bit category).