Implement kzalloc using ldv_kzalloc or ldv_zalloc
Removes one of the uses of __VERIFIER_nondet_pointer, which is
implemented using __VERIFIER_nondet_pointer. The size is known, thus
using ldv_(k)zalloc is perfectly safe, and also more consistent as it
will ensure zero-initialisation. The choice between ldv_zalloc and
ldv_kzalloc is made based on what is declared in each task. This is in
preparation of removing uses of __VERIFIER_nondet_pointer (see #767).

The changes were done using the following command:
```
for f in c/ldv-*/*.[ci]
do
perl -i -e \
  '$p = 1; $d = 0; $dk = 0; $l = "";
   while(<>) {
     if(/^void \*ldv_zalloc\(size_t size \) ;$/) {
       next if($d == 2);
       $d = 1;
     }
     if(/^void \*ldv_kzalloc\(size_t size , gfp_t flags \) ;$/) {
       next if($dk == 2);
       $dk = 1;
     }
     if(/^__inline static void \*kzalloc\(size_t size , gfp_t flags \) ?$/) {
       $p = 2;
     }
     if($p == 2) {
       if(/^  tmp = __VERIFIER_nondet_pointer\(\);/) {
         if($dk == 1) {
           print $l;
           print "  tmp = ldv_kzalloc(size, flags);\n";
         }
         else {
           if($d == 0) {
             print "void *ldv_zalloc(size_t size ) ;\n";
             $d = 2;
           }
           print $l;
           print "  tmp = ldv_zalloc(size);\n";
         }
         $l = "";
         $p = 1;
       }
       elsif(/^\}$/) {
         print $l;
         $l = "";
         print;
         $p = 1;
       }
       else { $l .= $_; }
     }
     else { print; }
   }' \
  $f
done
```