fix comparison of freed pointer.
According to ISO-C99 6.2.4(2),
the value of a pointer becomes indeterminate when
the object it points to reaches the end of its lifetime.

The undefined behavior comes from computing "ptr!=head" in this task,
because the memory behind "head" was freed before,
in the first loop iteration, when "free(ptr)" is called and ptr points to the same region as head.
So the value of head is indeterminate.

The fix uses a temporary variable for storing the corresponding value.

This problem is similar to #979.