Rename functions that collide with the C standard in part of AWS-C-Common
Part of #1043.
All these programs define some functions (at least abort)
that are also defined by the C standard,
but these identifiers are reserverd and so this is undefined
behavior.

This commit renames functions called "abort", "qsort", "memcpy",
"memmove", and "memset" if they have a body.
These were the only functions I found where this is necessary.

Checking which tasks are affected and the renaming was done
automatically, the code for this is in `makeall` of #1047.

This PR is a subset of the fixes of #1047
where it is relevant for SV-COMP'20.