list-ext-properties/list-ext.i is not memory safe
The last element appended to the list never has its .n field set. As
this is in memory allocated using malloc, any read access to this field
is undefined behaviour. This materialises as an invalid-free, because
p->n need not be null, and thus the iteration continues up to that point
(this access is undefined behaviour), and the pointer is passed to
free().