Remove undefined behaviour in list-ext-properties/list-ext.i
The last element appended to the list never had its .n field set. As this is in
memory allocated using malloc, any read access to this field is undefined
behaviour. This could materialise as an invalid-free, but really all bets are
off as the access is undefined behaviour. (This is _not_ an invalid
dereference.)