Initial preparation for making the repo REUSE-compliant
We want to use the best practices of the REUSE project for this
repository: https://reuse.software/

This will make maintenance of copyright and license easier.
For example, we sometimes move files between directories,
so it is better if the license is declared in each file,
instead of just per directory.
The REUSE project also provides a nice tool that we can use in CI
to check that every file has a proper license and copyright declaration.

This will also improve the situation for users of this repository,
e.g., for researchers who use it for experiments and redistribute
(parts of) the verification tasks in artifacts.
These will simply be able to take the files from here,
run "reuse download --all" to add all required licenses texts
and are done, no need to manually care about keeping license texts
and copyright notices.

This commits adds a few basic helper files,
adjusts the documentation,
and shows how this will work by adding headers
to the files in the project root directory.