Switch CI check for licenses
Instead of a check for a license file in each directory,
we now check for REUSE compliance.
Of course this will fail until we have converted all existing
declarations, so it is marked as allowed to fail for now.

Because of the inflexibility of the Travis configuration,
let's try GitHub actions for this in addition to GitLab CI.