Fix CI for new category NoDataRace-Main
This was merged via PR #1132 / commit 810e7d68. CI was already broken
by then so we did not see that this caused the CI to fail.
This commit will omit the checks of platform-dependent types for
NoDataRace-Main in the same fashion as it has already been done for
ConcurrencySafety-Main. Thanks to @hernanponcedeleon for his comment
in PR #1129, otherwise I would have needed a lot of time to figure
this out.