Fix verdicts for nla-digbench-scaling
The verdicts are adapted to represent the results of the first
tool executions on these tasks. This was expected to happen, verdicts
could still be wrong on a few tasks, the verifiers will hopefully find
this! Also this commit reverts some changes that were still left over
from reverting PR #1119, such that once the fixed PR comes we have a
consistent state.