Insert assert(0) into the body of reach_error
As discussed in the (virtual) community meeting on 2020-09-18,
unreach-call may be turned into plain assertion checking. As first step,
make the body of reach_error() `assert(0);`. For preprocessed files, the
result of assert(0) expansion is what we get on Ubuntu 18.04 with glibc
and GCC 7.5.0.

This change was previously made in 7341f59ba96, which did not properly handle
the case of reach_error not being defined in the main source file. This commit
now only handles openbsd-6.2 tasks, which define reach_error in a header file
and use files included in the repository for preprocessing. Other tasks will be
handled in a separate commit.

The changes in this commit were done manually, but do not yet affect the
actual tasks. Those will require automatic preprocessing, which will be
covered by the next commit.