Whitespace changes only: Re-preprocess c/openbsd-6.2
The introduction of reach_error() did not apply formatting otherwise
used in this directory. This results in `make modules` generating
additional (whitespace-only) changes.

The changes in this commit were obtained solely by running
```
make modules
```
in c/openbsd-6.2 with `clang-format` version 6.0.0.

The purpose of this whitespace-only commit is to reduce the size of
future changes to these benchmark tasks to a minimum.