Declare license of task-definition files and set files to be CC0
This was discussed in
https://github.com/sosy-lab/sv-benchmarks/pull/1099#issuecomment-698289859
It is questionable anyway whether these files contain any copyrightable
content. The set files are trivial and the task-definition files
were mostly auto-generated.

For this reason and because there are so many of these files,
we do not add a license header to each of them
but just add a central declaration in .reuse/dep5
(as allowed by the REUSE spec).