Final cleanup of license status for Juliet tasks
Fixes #1018

We also label the readme's with the same license as the sources
because they were submitted together by the respective contributors
and thus can be assumed to be under the same license.

The original C sources for Juliet tasks get a general license statement
in .reuse/dep5 because it is probably not worth it to change all of
these files.
For example, users of this repo who wish to redistribute parts of it
(e.g., in artifacts) will likely redistribute only the actual task
files, so for the task files the REUSE compatibility provides a benefit.