Update task defs with properties coverage-branches and coverage-error-call
Over the course of the last year, a lot of tasks were fixed and a a lot of tasks were added.
Most of the time, people did not consider the coverage-* properties for these tasks, so let's add
them now.

Properties are added according to the criteria stated in !842:

    * Compiles, if __VERIFIER_X method definitions are provided,
    * No termination property or expected verdict for termination is true, and
    * Some _VERIFIER_nondet* method exists that is not only declared, but used in at least one other location

Compliance is checked and properties are added with the scripts found at
https://gist.github.com/lembergerth/1a1bf782931fb16af0d9e4bc1085a737 .

Ability to compile is checked with `bin/prtest --compile-only`,
PRTest version
https://gitlab.com/sosy-lab/software/prtest/-/tree/cc9cc249