change labelling for uthash tasks.
This is the same change as the previous commit, but different tasks.

Those tasks all contain the same behaviour:
They allocate 10 users, but never free them.
The allocated memory is only reachable via local variables,
and so there is a memtrack violation when exiting the main function.

Found by Symbiotic.