seq-* benchmarks: Remove use of function without definition
These benchmarks include a definition of a `nodes` variable that is
initialized with pointers to functions that are declared, but not
defined. This is undefined behaviour, and makes linking (and thus
running) those benchmarks impossible.

The changes were done using the following command:
```
for f in c/seq-*/*.c
do
perl -i -e \
  '$i = 0;
   while(<>) {
     if($i == 0 && /^void\s*\(\*nodes\[\d+\]\)\(void\)\s*=\s*\{/) {
       $i = 1 unless(/\}\s*;/);
       next;
     }
     elsif($i == 1) {
       $i = 0 if(/\}\s*;/);
       next;
     }
     print;
   }' $f
done
```