Fix verdict of value-bounded freire1 tasks to 'true'
the assertion in these tasks checks that,
for int r and double a,
r^2 - r - a + 2 * (a / 2 - sum(range(0, r))) == 0
(for range-end exclusive)

If there's no imprecisions due to floating-point arithmetics,
this equality is correct.

In the value-bounded tasks, a is restricted to 0 <= a <= 100.
In this value-range, no imprecision is introduced.
Thus, the tasks fulfill property unreach-call.